backend/pyproject.toml
[tool.poetry]
name = "University-Nebraska-Tutor-Portal"
version = "0.0.1"
description = "Application to facilitate communication between students and tutors by allowing students to create, manage, and track tickets for tutoring assistance."
authors = ["Nolan Gregory <nolangregory@unomaha.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
django = "^4.2.3"
python-dotenv = "^1.0.0"
pytest = "^7.4.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
websockets = "^11.0.3"
channels = {extras = ["daphne"], version = "^4.0.0"}
pandas = "^2.1.1"
django-extensions = "^3.2.3"
sphinx = "^7.2.6"
sphinxawesome-theme = "^4.1.0"
furo = "^2023.9.10"
sphinx-copybutton = "^0.5.2"
myst-parser = "^2.0.0"
pyparsing = "^3.1.1"
pydot = "^1.4.2"

[tool.poetry.group.dev.dependencies]
bruhcolor = "^0.0.56"
pylint = "^3.0.1"
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
mypy = "^1.4.1"
ruff = "^0.0.282"
yapf = "^0.40.1"
autoflake = "^2.2.0"
pytest-django = "^4.5.2"
pip-audit = "^2.6.1"
autobahn = "^23.6.2"
bandit = "^1.7.5"
django-stubs = "^4.2.6"
coverage = "^7.3.2"
pytest-cov = "^4.1.0"
autopep8 = "^2.0.4"
faker = "^19.12.0"
werkzeug = "^3.0.1"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
extend-exclude = '''
(
    | settings.py
)
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "base.settings"
python_files = ["test_*.py", "*_test.py"]

[tool.ruff]
exclude = [".venv", "tests/", "scripts"]
line-length = 100

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402"]
"settings.py" = ["E501"]

[tool.mypy]
ignore_missing_imports = "True"
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "base.settings"

[tool.autoflake]
imports = ["django", "requests", "urllib3"]

[tool.bandit]
exclude_dirs = [".venv", "tests"]
skips = ["B101"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
